+++ Back/Banco-de-Talentos/src/main/java/com/fractal/bancodetalentos/model/response/TechnicalAbilitiesResp.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -12,5 +14,5 @@ import lombok.Setter;
 public class TechnicalAbilitiesResp {
     private Integer idTechnicalAbility;
     private String name;
-    private Integer years;
+    private BigDecimal years;
 }
